reafactor custom_hook
